backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
llama-cpp-python
